refacor: reorganize notification and outline symbol providers
