+++ src/providers/notification/notificationAction.ts
@@ -1,5 +1,5 @@
 import * as vscode from 'vscode';
-import { outputChannel } from './extension';
+import { outputChannel } from '../../extension';
 
 // Unused ideas
 // https://code.visualstudio.com/api/extension-guides/webview
