+++ src/providers/notification/registerUpdateInfo.test.ts
