+++ src/providers/outline/registerOutlineSymbol.ts
