+++ src/providers/outline/symbolsInfo.test.ts
