+++ src/providers/outline/symbolsInfo.ts
